+++ src/NerdStore.Catalogo.Domain/Categoria.cs
@@ -1,0 +1,27 @@
+﻿using NerdStore.Core;
+
+namespace NerdStore.Catalogo.Domain;
+
+public class Categoria : Entity
+{
+    public string Nome { get; private set; }
+    public string Codigo { get; private set; }
+
+    public Categoria(
+        string nome,
+        string codigo)
+    {
+        Nome = nome;
+        Codigo = codigo;
+    }
+
+    public override string ToString()
+    {
+        return $"{Nome} - {Codigo}";
+    }
+
+    public void Validar()
+    {
+        
+    }
+}
