+++ src/NerdStore.Catalogo.Domain/Produto.cs
@@ -67,27 +67,3 @@ public class Produto : Entity, IAggregateRoot
         
     }
 }
-
-public class Categoria : Entity
-{
-    public string Nome { get; private set; }
-    public string Codigo { get; private set; }
-
-    public Categoria(
-        string nome,
-        string codigo)
-    {
-        Nome = nome;
-        Codigo = codigo;
-    }
-
-    public override string ToString()
-    {
-        return $"{Nome} - {Codigo}";
-    }
-
-    public void Validar()
-    {
-        
-    }
-}
